+++ src/equal.cpp
@@ -1,0 +1,124 @@
+#include "pointer.hpp"
+#include <HepMC3/GenEvent.h>
+#include <HepMC3/GenHeavyIon.h>
+#include <HepMC3/GenParticle.h>
+#include <HepMC3/GenRunInfo.h>
+#include <HepMC3/GenVertex.h>
+
+namespace HepMC3 {
+
+template <class T>
+bool operator!=(const T& a, const T& b) {
+  return !operator==(a, b);
+}
+
+// equality comparions used by unit tests
+bool is_close(const FourVector& a, const FourVector& b, double rel_eps = 1e-7) {
+  auto is_close = [rel_eps](double a, double b) { return std::abs(a - b) < rel_eps; };
+  return is_close(a.x(), b.x()) && is_close(a.y(), b.y()) && is_close(a.z(), b.z()) &&
+         is_close(a.t(), b.t());
+}
+
+// compares all real qualities of two particles, but ignores the .id() field
+bool operator==(const GenParticle& a, const GenParticle& b) {
+  return a.pid() == b.pid() && a.status() == b.status() &&
+         is_close(a.momentum(), b.momentum());
+}
+
+// compares all real qualities of both particle sets,
+// but ignores the .id() fields and the particle order
+bool equal_particle_sets(const std::vector<ConstGenParticlePtr>& a,
+                         const std::vector<ConstGenParticlePtr>& b) {
+  if (a.size() != b.size()) return false;
+  auto unmatched = b;
+  for (auto&& ai : a) {
+    auto it = std::find_if(unmatched.begin(), unmatched.end(),
+                           [ai](ConstGenParticlePtr x) { return *ai == *x; });
+    if (it == unmatched.end()) return false;
+    unmatched.erase(it);
+  }
+  return unmatched.empty();
+}
+
+// compares all real qualities of two vertices, but ignores the .id() field
+bool operator==(const GenVertex& a, const GenVertex& b) {
+  return a.status() == b.status() && is_close(a.position(), b.position()) &&
+         equal_particle_sets(a.particles_in(), b.particles_in()) &&
+         equal_particle_sets(a.particles_out(), b.particles_out());
+}
+
+// compares all real qualities of both vertex sets,
+// but ignores the .id() fields and the vertex order
+bool equal_vertex_sets(const std::vector<ConstGenVertexPtr>& a,
+                       const std::vector<ConstGenVertexPtr>& b) {
+  if (a.size() != b.size()) return false;
+  auto unmatched = b;
+  for (auto&& ai : a) {
+    auto it = std::find_if(unmatched.begin(), unmatched.end(),
+                           [ai](ConstGenVertexPtr x) { return *ai == *x; });
+    if (it == unmatched.end()) return false;
+    unmatched.erase(it);
+  }
+  return unmatched.empty();
+}
+
+bool operator==(const GenRunInfo::ToolInfo& a, const GenRunInfo::ToolInfo& b) {
+  return a.name == b.name && a.version == b.version && a.description == b.description;
+}
+
+bool operator==(const GenRunInfo& a, const GenRunInfo& b) {
+  const auto a_attr = a.attributes();
+  const auto b_attr = b.attributes();
+  return a.tools().size() == b.tools().size() &&
+         a.weight_names().size() == b.weight_names().size() &&
+         a_attr.size() == b_attr.size() &&
+         std::equal(a.tools().begin(), a.tools().end(), b.tools().begin()) &&
+         std::equal(a.weight_names().begin(), a.weight_names().end(),
+                    b.weight_names().begin()) &&
+         std::equal(a_attr.begin(), a_attr.end(), b_attr.begin(),
+                    [](const std::pair<std::string, AttributePtr>& a,
+                       const std::pair<std::string, AttributePtr>& b) {
+                      if (a.first != b.first) return false;
+                      if (bool(a.second) != bool(b.second)) return false;
+                      if (!a.second) return true;
+                      std::string sa, sb;
+                      a.second->to_string(sa);
+                      b.second->to_string(sb);
+                      return sa == sb;
+                    });
+}
+
+bool operator==(const GenHeavyIon& a, const GenHeavyIon& b) {
+  return a.Ncoll_hard == b.Ncoll_hard && a.Npart_proj == b.Npart_proj &&
+         a.Npart_targ == b.Npart_targ && a.Ncoll == b.Ncoll &&
+         a.N_Nwounded_collisions == b.N_Nwounded_collisions &&
+         a.Nwounded_N_collisions == b.Nwounded_N_collisions &&
+         a.Nwounded_Nwounded_collisions == b.Nwounded_Nwounded_collisions &&
+         a.impact_parameter == b.impact_parameter &&
+         a.event_plane_angle == b.event_plane_angle &&
+         a.sigma_inel_NN == b.sigma_inel_NN && a.centrality == b.centrality &&
+         a.user_cent_estimate == b.user_cent_estimate;
+}
+
+bool operator==(const GenEvent& a, const GenEvent& b) {
+  // incomplete:
+  // missing comparison of GenHeavyIon, GenPdfInfo, GenCrossSection
+
+  if (a.event_number() != b.event_number() || a.momentum_unit() != b.momentum_unit() ||
+      a.length_unit() != b.length_unit())
+    return false;
+
+  // run_info may be missing
+  if (a.run_info() && b.run_info()) {
+    if (!(*a.run_info() == *b.run_info())) return false;
+  } else if (!a.run_info() && b.run_info()) {
+    if (*b.run_info() != GenRunInfo()) return false;
+  } else if (a.run_info() && !b.run_info()) {
+    if (*a.run_info() != GenRunInfo()) return false;
+  }
+
+  // if all vertices compare equal, then also all particles are equal
+  return equal_vertex_sets(a.vertices(), b.vertices());
+}
+
+} // namespace HepMC3
